feat(thread.{h,-c},-process.c,-syscall.c,-exception.c,-debug.c): implemented wait syscall
and process_wait; introduced THREAD_DEAD; now when threads call thread_exit their resources aren't actually deallocated. Instead, they are marked as DEAD so that later on their parent can retrieve their exit status. A thread's resources are deallocated if and only if it's DEAD and its parent is DEAD as well or if the thread is passed as an argument to process_wait